backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1
psycopg2-binary==2.9.10
pydantic==2.10.4
pydantic-settings==2.7.1
python-multipart==0.0.18
ldap3==2.9.1
PyJWT==2.9.0
